+++ src/models/schemaRegistry.ts
@@ -1,6 +1,7 @@
 import { Data, type Require as Enforced } from "dataclass";
 import * as vscode from "vscode";
 import { CCLOUD_CONNECTION_ID, IconNames } from "../constants";
+import { CustomMarkdownString } from "./main";
 
 // Main class representing CCloud Schema Registry clusters, matching key/value pairs returned
 // by the `confluent schema-registry cluster describe` command.
@@ -27,20 +28,38 @@ export class SchemaRegistryClusterTreeItem extends vscode.TreeItem {
     const label = "Schema Registry";
     super(label, vscode.TreeItemCollapsibleState.None);
 
+    // internal properties
     this.resource = resource;
-    this.description = this.resource.id;
-
-    // TODO: update based on product+design feedback
-    this.tooltip = JSON.stringify(this.resource, null, 2);
-
     this.contextValue = "ccloud-schema-registry-cluster";
 
+    // user-facing properties
+    this.description = this.resource.id;
     this.iconPath = new vscode.ThemeIcon(IconNames.SCHEMA_REGISTRY);
+    this.tooltip = createSchemaRegistryClusterTooltip(this.resource);
 
+    // set primary click action to select this cluster as the current one, focusing it in the Schemas view
     this.command = {
       command: "confluent.resources.schema-registry.select",
       title: "Set Current Schema Registry Cluster",
       arguments: [this.resource],
     };
   }
+}
+
+function createSchemaRegistryClusterTooltip(
+  resource: SchemaRegistryCluster,
+): vscode.MarkdownString {
+  // TODO(shoup) update for local SR once available
+  const tooltip = new CustomMarkdownString()
+    .appendMarkdown(`#### $(${IconNames.SCHEMA_REGISTRY}) Confluent Cloud Schema Registry Cluster`)
+    .appendMarkdown("\n\n---\n\n")
+    .appendMarkdown(`ID: \`${resource.id}\`\n\n`)
+    .appendMarkdown(`Provider: \`${resource.provider}\`\n\n`)
+    .appendMarkdown(`Region: \`${resource.region}\`\n\n`)
+    .appendMarkdown(`URI: \`${resource.uri}\``)
+    .appendMarkdown("\n\n---\n\n")
+    .appendMarkdown(
+      `[$(${IconNames.CONFLUENT_LOGO}) Open in Confluent Cloud](${resource.ccloudUrl})`,
+    );
+  return tooltip;
 }