+++ src/models/topic.ts
@@ -2,6 +2,7 @@ import { Data, type Require as Enforced } from "dataclass";
 import * as vscode from "vscode";
 import { KafkaTopicOperation } from "../authz/types";
 import { CCLOUD_CONNECTION_ID, IconNames, LOCAL_CONNECTION_ID } from "../constants";
+import { CustomMarkdownString } from "./main";
 /** Main class representing Kafka topic */
 export class KafkaTopic extends Data {
   name!: Enforced<string>;
@@ -53,48 +54,81 @@ export class KafkaTopicTreeItem extends vscode.TreeItem {
   resource: KafkaTopic;
 
   constructor(resource: KafkaTopic) {
-    // these will always have at least Configurations to expand
-    super(resource.name, vscode.TreeItemCollapsibleState.Collapsed);
+    super(resource.name);
 
-    const tooltipBuf: string[] = [];
-
+    // internal properties
     this.resource = resource;
-
     this.contextValue = resource.isLocalTopic() ? "local-kafka-topic" : "ccloud-kafka-topic";
+    if (this.resource.hasSchema) {
+      this.contextValue += "-with-schema";
+      this.collapsibleState = vscode.TreeItemCollapsibleState.Collapsed;
+    } else {
+      this.collapsibleState = vscode.TreeItemCollapsibleState.None;
+    }
 
-    // Convert 'read' and 'delete' operations to context flags to hang context-sensitive commands off of.
-    // (these are interesting because we invite specific actions based on their presence)
+    // user-facing properties
+    // this.description = "";  // TBD
+    this.iconPath = resource.hasSchema
+      ? new vscode.ThemeIcon(IconNames.TOPIC)
+      : new vscode.ThemeIcon(
+          IconNames.TOPIC_WITHOUT_SCHEMA,
+          new vscode.ThemeColor("problemsWarningIcon.foreground"),
+        );
+
+    const missingAuthz: KafkaTopicOperation[] = this.checkMissingAuthorizedOperations(resource);
+    this.tooltip = createKafkaTopicTooltip(this.resource, missingAuthz);
+  }
+
+  checkMissingAuthorizedOperations(resource: KafkaTopic): KafkaTopicOperation[] {
+    // operations we support via view/item actions that require authorization
     const interestingAuthz: KafkaTopicOperation[] = ["READ", "DELETE"];
-    const missingAuthz: KafkaTopicOperation[] = [];
+
     for (const op of interestingAuthz) {
       if (resource.operations.includes(op)) {
         // Convert to "authzRead", "authzDelete", etc. for context flags to hang context-sensitive commands off of (see package.json)
         const operationTitleCase = op.toLowerCase().replace(/^\w/, (c) => c.toUpperCase());
         this.contextValue += `-authz${operationTitleCase}`;
-      } else {
-        missingAuthz.push(op);
       }
     }
 
-    if (missingAuthz.length > 0) {
-      tooltipBuf.push("⚠️ Missing authorization for the following actions:");
-      missingAuthz.forEach((op) => tooltipBuf.push(` * ${op}`));
-      tooltipBuf.push("");
-    }
-
-    if (resource.hasSchema) {
-      this.iconPath = new vscode.ThemeIcon(IconNames.TOPIC);
-      this.contextValue = `${this.contextValue}-with-schema`;
-    } else {
-      this.collapsibleState = vscode.TreeItemCollapsibleState.None;
-      this.iconPath = new vscode.ThemeIcon(
-        IconNames.TOPIC_WITHOUT_SCHEMA,
-        new vscode.ThemeColor("problemsWarningIcon.foreground"),
-      );
-      tooltipBuf.push("⚠️ No schema(s) found for topic.");
-      tooltipBuf.push("");
-    }
-
-    this.tooltip = tooltipBuf.join("\n");
+    return interestingAuthz.filter((op) => !resource.operations.includes(op));
   }
+}
+
+function createKafkaTopicTooltip(
+  resource: KafkaTopic,
+  missingAuthz: KafkaTopicOperation[],
+): vscode.MarkdownString {
+  const tooltip = new CustomMarkdownString();
+  const iconName = resource.hasSchema ? IconNames.TOPIC : IconNames.TOPIC_WITHOUT_SCHEMA;
+  tooltip
+    .appendMarkdown(`#### $(${iconName}) Kafka Topic`)
+    .appendMarkdown("\n\n---\n\n")
+    .appendMarkdown(`Name: \`${resource.name}\`\n\n`)
+    .appendMarkdown(`Replication Factor: \`${resource.replication_factor}\`\n\n`)
+    .appendMarkdown(`Partition Count: \`${resource.partition_count}\`\n\n`)
+    .appendMarkdown(`Internal: \`${resource.is_internal}\`\n\n`);
+
+  if (!resource.hasSchema) {
+    tooltip
+      .appendMarkdown("---\n\n")
+      .appendMarkdown("$(warning) No schema(s) found for topic.\n\n");
+  }
+
+  // list any missing authorized operations
+  if (missingAuthz.length > 0) {
+    tooltip
+      .appendMarkdown("---\n\n")
+      .appendMarkdown("$(warning) Missing authorization for the following actions:\n\n");
+    missingAuthz.forEach((op) => tooltip.appendMarkdown(` - ${op}\n`));
+  }
+
+  if (!resource.isLocalTopic()) {
+    tooltip.appendMarkdown("---\n\n");
+    tooltip.appendMarkdown(
+      `[$(${IconNames.CONFLUENT_LOGO}) Open in Confluent Cloud](${resource.ccloudUrl})`,
+    );
+  }
+
+  return tooltip;
 }