+++ src/sidecar/connections.ts
@@ -1,11 +1,7 @@
 import { getSidecar } from ".";
 import { Connection, ConnectionsResourceApi, ResponseError } from "../clients/sidecar";
 import { CCLOUD_CONNECTION_ID, CCLOUD_CONNECTION_SPEC } from "../constants";
-import {
-  currentCCloudEnvironmentChanged,
-  currentKafkaClusterChanged,
-  currentSchemaRegistryChanged,
-} from "../emitters";
+import { currentKafkaClusterChanged, currentSchemaRegistryChanged } from "../emitters";
 import { Logger } from "../logging";
 import { getResourceManager } from "../storage/resourceManager";
 
@@ -56,18 +52,12 @@ export async function deleteCCloudConnection(): Promise<void> {
 }
 
 export async function clearCurrentCCloudResources() {
-  // if the current connection changes or is deleted, we need to unset any associated CCloud resources
+  // if the current connection changes or is deleted, we need to clear any associated CCloud resources
   // that may have depended on it:
   // - delete the extension state references to make sure they can't be used
   // - fire events to update things like the Topics view, Schemas view, etc.
-  const resourceManager = getResourceManager();
   logger.warn("clearing current CCloud resources from extension state");
-  await Promise.all([
-    resourceManager.deleteCCloudEnvironments(),
-    resourceManager.deleteCCloudKafkaClusters(),
-    resourceManager.deleteCCloudSchemaRegistryClusters(),
-  ]);
-  currentCCloudEnvironmentChanged.fire(null);
+  await getResourceManager().deleteCCloudResources();
   currentKafkaClusterChanged.fire(null);
   currentSchemaRegistryChanged.fire(null);
 }